refactor: give the shared policies one home each
Five changes, each findable without a metric, plus the honest accounting for
two that were not worth making.

A next-page href is untrusted data, and validating one existed three times with
three behaviors. ogc/engine.py raised a bare RuntimeError on a cross-host link,
did not resolve a relative href (it handed the unresolved reference back as the
pagination cursor, so the follow-up request failed instead of paging), and fell
open by returning the raw href when httpx.URL() rejected it. ratings resolved,
required an exact host, stripped userinfo. wateruse resolved, accepted a host
alias set, rewrote to the canonical host, dropped the port too. One policy now
lives in transport/links.py and the walks pass in what genuinely differs -- the
acceptable hosts and whether an accepted host is rewritten. Both engine bugs
are fixed by the merge. Its fall-open branch was reachable only from tests
whose mock response had no real URL, so those mocks got one rather than the
security check keeping a hole to accommodate them. Each walk keeps the
exception type it already raised: retyping engine's RuntimeError to
DataRetrievalError is a released behavior change to make deliberately, so
resolve_next_url takes the type and engine passes RuntimeError.

parse_retry_after moved from transport/retry.py, which never called it, to
exceptions.py, beside the DataRetrievalError.retry_after field it exists to
produce. Its two callers each use it in one expression, both feeding
error_for_status. Three stdlib imports went dead in retry.py on the way out --
evidence it shared nothing with its host module. ogc/errors.py, whose job is
mapping a status to a typed exception, no longer reaches through the retry
policy (and thence interruptions and pandas) to parse a header.

utils.py split along its real seam. Half was pandas column munging over codes;
half was "GET a USGS URL, map the status, retry it" over exceptions and
transport. They shared a filename and nothing else -- nldi, streamstats and
nwis took only the HTTP half, waterdata.samples only the munging. The query
path is now _querying.py, private for the same reason _ambient and
_response_metadata are: the names are public even though the module is not, so
utils.query and utils.to_str stay the documented paths (__module__ preserved,
so the API reference page is unchanged).

Also: waterdata profile validation moved into types.py, next to the three
tables it exists to validate against, which was waterdata/utils.py's only
reason to import types; nwis.get_dv and get_iv now share one body, which
encodes the non-obvious contract that the caller-facing args are aliases that
lose to explicit waterservices kwargs; nldi's repeated "fetch, honor as_json"
tail became _query_features; and utils.py lost four compatibility aliases for
names that never existed at that path (git show v1.2.0 confirms) and were
referenced nowhere. One of the five candidates for removal, HTTPX_ASYNC_DEFAULTS,
turned out not to be dead -- a test imports it -- so it was repointed at
transport.http rather than deleted blind.

pyscn: composite 81 -> 82. Dependencies 75 -> 80 (max depth 9 -> 8, total
dependencies 144 -> 149 across two new modules), architecture 81 -> 82
(weighted violations 39 -> 37; dataretrieval.utils drops out of ERROR),
duplication holds at 70 with clone fragments 26 -> 25 and 9.25% -> 8.90% --
real, but the 75 band starts near 7.3%. Complexity, dead code, coupling and
cohesion unchanged at 95/100/100/100.

Two candidates were measured and declined. Splitting waterdata/utils.py three
ways (policy/args/retrieval) scores +0.65 architecture, but its weighted
violations are unchanged -- one 5-point error traded for five 1-point warnings
-- so the entire gain is denominator drift from the extra modules, and the new
warnings exist because "args" and "retrieval" have real names where "utils" was
filtered out of every consumer's responsibility set. Only the part that stands
without the number was taken. Extracting the cause-walking predicates out of
interruptions.py into a leaf earns zero dependency points (interruptions gains
the edge retry loses) and costs a new module plus a layers edit on the module
that owns the public resume contract; it is worth doing when interruptions is
being refactored for its own reasons, not for this.

Pruning the dead utils aliases is scored slightly negative (architecture
compliance 80.60 -> 80.50: utils goes from 12 responsibilities to 11 but stays
an ERROR while TotalRules falls, so the ratio drops). Kept anyway -- a
compatibility promise for a name that never shipped is debt, not surface.

Tests 779 passed / 12 deselected throughout, mypy clean (53 -> 55 files with
the two new modules), xenon, complexipy, import-linter and pre-commit green.
