+++ .importlinter
@@ -24,6 +24,7 @@ layers =
     ngwmn | nldi | nwis | streamstats | waterdata | wateruse | wqp
     ogc
     utils
+    _querying
     transport
     progress
     _ambient | _response_metadata | codes | combining | interruptions | rdb
@@ -89,12 +90,14 @@ type = forbidden
 source_modules =
     dataretrieval.ogc
 forbidden_modules =
+    dataretrieval._querying
     dataretrieval.utils
 
 [importlinter:contract:nwis-quarantine]
 name = Deprecated NWIS has no dependents (ADR 0005)
 type = forbidden
 source_modules =
+    dataretrieval._querying
     dataretrieval.codes
     dataretrieval.combining
     dataretrieval.credentials
