+++ dataretrieval/waterdata/ratings.py
@@ -17,7 +17,7 @@ from typing import Any, Literal, get_args
 import httpx
 import pandas as pd
 
-from dataretrieval.credentials import WATERDATA_BASE_URL, without_embedded_credentials
+from dataretrieval.credentials import WATERDATA_BASE_URL
 from dataretrieval.exceptions import DataRetrievalError
 from dataretrieval.ogc.dates import _DURATION_RE, _format_api_dates
 from dataretrieval.ogc.errors import _raise_for_non_200
@@ -33,6 +33,7 @@ from dataretrieval.transport.http import (
 from dataretrieval.transport.http import (
     get as _get,
 )
+from dataretrieval.transport.links import resolve_next_url
 
 __all__ = ["get_ratings"]
 
@@ -245,28 +246,6 @@ def _search(
     one page isn't silently truncated.
     """
 
-    def _checked_next_url(href: str, current: httpx.URL) -> str:
-        """Resolve a server-supplied ``next`` href into a URL safe to request."""
-        try:
-            target = httpx.URL(href)
-        except (httpx.InvalidURL, TypeError) as exc:
-            raise DataRetrievalError(
-                f"The ratings service returned an unusable next-page link: "
-                f"{href!r}. The page walk cannot continue; report this if it "
-                f"persists."
-            ) from exc
-        if not target.is_absolute_url:
-            target = current.join(target)
-        if target.host != current.host:
-            raise DataRetrievalError(
-                f"Refusing to follow a ratings next-page link pointing at "
-                f"{target.host} rather than {current.host}. Following it would "
-                f"send this request, and any credentials on it, to a host you "
-                f"did not ask for. Retrying will not help; report this if it "
-                f"persists."
-            )
-        return str(without_embedded_credentials(target))
-
     query_params: dict[str, Any] = {"limit": min(limit, 10000)}
     if filter_str is not None:
         query_params["filter"] = filter_str
@@ -303,7 +282,11 @@ def _search(
         # carry this request's API key off the authorized host, and one carrying
         # ``user:pass@`` would mint an ``Authorization: Basic`` header the caller
         # never configured.
-        url = None if href is None else _checked_next_url(href, response.url)
+        url = (
+            None
+            if href is None
+            else resolve_next_url(href, response, service="ratings")
+        )
         params = None
     return features
