+++ dataretrieval/waterdata/samples.py
@@ -18,6 +18,7 @@ from urllib.parse import quote
 import httpx
 import pandas as pd
 
+from dataretrieval._querying import to_str
 from dataretrieval._response_metadata import BaseMetadata
 from dataretrieval.ogc.errors import _raise_for_non_200
 from dataretrieval.transport.http import (
@@ -29,16 +30,16 @@ from dataretrieval.transport.http import (
 from dataretrieval.transport.http import (
     get as _get,
 )
-from dataretrieval.utils import _attach_datetime_columns, to_str
+from dataretrieval.utils import _attach_datetime_columns
 from dataretrieval.waterdata.types import (
     CODE_SERVICES,
     PROFILES,
     SERVICES,
+    _check_profiles,
 )
 from dataretrieval.waterdata.utils import (
     SAMPLES_URL,
     _accept_legacy_kwargs,
-    _check_profiles,
     _get_args,
 )
