+++ docs/source/architecture/index.rst
@@ -149,11 +149,21 @@ Shared components
     import and resolves to this same class.
 
 ``dataretrieval.utils``
-    Data-shaping helpers, legacy request composition, and compatibility imports
-    for names that historically lived here (including ``Ambient`` and
-    ``BaseMetadata``, so their original import paths keep working). OGC does not
-    depend on this mixed legacy module; by default, do not add new
-    service-specific behavior there.
+    Data-shaping helpers, plus compatibility imports for names that
+    historically lived here (including ``Ambient``, ``BaseMetadata``, ``query``
+    and ``to_str``, so their original import paths keep working). OGC does not
+    depend on this legacy module; by default, do not add new service-specific
+    behavior there.
+
+``dataretrieval._querying``
+    The one-shot HTTP query path the single-request adapters (``nwis``,
+    ``wqp``, ``nldi``, ``streamstats``, ``wateruse``) use: compose the URL, send
+    it, map the status, retry a transient. It left ``utils`` because the two
+    halves shared only a filename -- this one depends on ``exceptions`` and
+    ``transport``, the shaping half on ``codes`` and pandas, and no caller
+    wanted both. The implementation module is private; the established public
+    function paths remain ``dataretrieval.utils.query`` and
+    ``dataretrieval.utils.to_str``.
 
 ``dataretrieval.codes`` and ``dataretrieval.rdb``
     State/time-zone code conversion and RDB parsing leaves.
@@ -304,8 +314,10 @@ This view records categories and representative locations of debt.
 ``.importlinter`` is authoritative for exact current dependency allowlists.
 
 - ``ogc/engine.py`` retains compatibility wrappers alongside OGC orchestration.
-- ``utils.py`` combines metadata, shaping, ambient configuration, legacy
-  request composition, and transport compatibility imports.
+- ``utils.py`` combines shaping with compatibility imports for metadata,
+  ambient configuration, transport, and the query path.
+- ``waterdata/utils.py`` combines endpoint constants, argument normalization,
+  and the OGC engine wrappers.
 
 These are documented so guardrails distinguish accepted current dependencies
 from new erosion. They should be removed through small, test-protected changes,
